rpc: Avoid using zero-copy interface of output_stream
That interface is currently lacking batching, so every write is passed
directly to the underlying data sink. This results in more smaller
packets to be sent, higher system time, more interrupts and thus
higher softirq time.

This led to a 30% regression in throughput of ScyllaDB in a CPU-bound
workload on EC2. Refs scylladb/scylla#1786.

Fix by reverting to the copying write, which buffers data and can
benefit from flush() batching.

Introduced in commit 016ced1.

Message-Id: <1477650683-7688-1-git-send-email-tgrabiec@scylladb.com>