core: Introduce weak_ptr<>
Allows obtaining a non-owning reference to the object. A live
weak_ptr<> object doesn't prevent the referenced object form being
destroyed.

The underlying pointer held by weak_ptr<> is valid as long as the
referenced object is alive.  When the object dies, all weak_ptr
objects associated with it are emptied.

The user of weak_ptr can check if it still holds a valid pointer like
this:

     if (ptr) ptr->do_something();

A weak reference is obtained like this:

     class X : public weakly_referencable<X> {};
     auto x = std::make_unique<X>();
     weak_ptr<X> x_ref = x->weak_from_this();

Message-Id: <1479224002-945-1-git-send-email-tgrabiec@scylladb.com>