metrics: Do not detroy the metrics layer if it is been used
This patch prevent the metrics layer to be destroyed if there are still
registered metrics.

The metrics impl will be created by the get_impl() function as a
shared_ptr (shared_lw_ptr requires class definition and cause cyclic
dependencies problem).

Each registered_metric object will hold a ptr to the implementation, so
while there are still registered metrics the implementation will not be
destroyed.

Signed-off-by: Amnon Heiman <amnon@scylladb.com>
Message-Id: <1479830410-29369-2-git-send-email-amnon@scylladb.com>