thread: check need_preempt for threads in a scheduling group as well
When we check for should_yield(), we should check for need_preempt() as
well. Otherwise threads relying on should_yield() only, that won't call
get() often may run for more than the task quota.

We should also take this into account when we actually yield: threads in a
scheduling group will only yield when their task quota is fully used, but if
there is need for preempt we should at least break the current execution.

Signed-off-by: Glauber Costa <glauber@scylladb.com>
Message-Id: <1480615139-8688-1-git-send-email-glauber@scylladb.com>