metrics: Create the metrics group in its constructor
This patch adds a wrapper to the unique_ptr that holds the metrics, so
it will create the metrics_group in its constructor and expose its
functionality.

This save the users of the metrics layer the need to instantiate the
parameters and add the add_group method to the metric_groups.

Signed-off-by: Amnon Heiman <amnon@scylladb.com>