fstream: Auto-close data_sink and data_source
In exceptional paths, it is hard to ensure that close() is always
called and waited for before the stream object goes out of scope. Not
doing so will currently result in undefined behavior, which can crash
the whole node. Refs scylladb/scylla#990, scylladb/scylla#1076.

After this change, the objects will be closed in the background when
they were not closed explicitly before they go out of scope.

It is still required to call close() explcitily in order to be
notified about errors, so non-exceptional paths should do it.

It is still required that the object stays alive as long as any
operation started on it (e.g. put, get, close) is in progress.

Message-Id: <1482428953-22888-1-git-send-email-tgrabiec@scylladb.com>