file: add dup() support
Experience has shown in Scylla that we sometimes need to open large
numbers of files on all shards.  This causes an explosion in the number
of file descriptors, requiring ulimits to be raised to ridiculous levels.

To fix, introduce file::dup(), which produces a seastar::file_handle
object that can be transported across shards, and there converted to
a new file object that refers to the same underlying file.

The interface was crafted to support non-posix files (e.g. a userspace
filesystem, or layered files like Scylla's checked_file_impl), so the
handle supports arbitrary data that may or may not need reference counting.

Message-Id: <20161225132541.31828-1-avi@scylladb.com>