metrics: Add label support
Using labels is useful when there is a metric that is reported by
multiple objects, but some of the time you need to aggregate it.
For example, when you have multiple queues per shard.

This patch adds a label and label_instance objects.

label_instance is added to a metrics when it is created. Multiple labels
can be assign to the metrics.

Typically multiple objects will add the same label (like column family).
In those cases, the label helper class will be used to create the
label_instance objects.

The metrics creation function where modified to accept labels.

The labels parameter where added in a way that would have broken the
function API, but, at this point there are almost no places in the code
that uses the default parameters, so it's better to fix this point in
the code and get a cleaner API.

Signed-off-by: Amnon Heiman <amnon@scylladb.com>