metrics: metrics keep the value map while unregistering
The register metrics holds a shared_ptr to the metrics layer, so that as
long as the registration is kept, the layer will not be destroyed.

This patch change the unregistration method, that during that
unregistration the method would hold the reference so it will be kept
alive until the end of the method.

Message-Id: <1483535892-14846-2-git-send-email-amnon@scylladb.com>