scripts: posix_net_conf.sh: use RPS queues number as FP IRQs number estimate
Number of RPS queues equals to the number of HW Rx queues and each HW Rx
queue has a separate IRQ. Theoretically a NIC may configure a separate
IRQ for HW Tx queue but this is going to be a waste of IRQ anyway and is
not an interesting case. So, when we want to get the number of "hard working"
IRQs, the number of RPS queues is exactly the number we need.

If RPS feature is not enabled, use number of NIC's IRQs as an estimate for
the number of Rx queues.

Signed-off-by: Vlad Zolotarov <vladz@scylladb.com>