tls.cc: Fix shutdown_input/output to conform with expected socket behaviour
Socket shutdown is used (and thus implicitly defined work as) asynchronously
with pending read/write operations. Esp. the rpc code uses shutdown to wake
itself up on stop.

The layered and slightly akward IO flow of the tls layer did not handle
this, and we ended up orphaning a waiting read promise.

Modify to wrap waits in a semaphore we can use to detect simultaneous
shutdown/handshake and actual data read/write, and if needed gate
through.

Message-Id: <1484588637-10196-1-git-send-email-calle@scylladb.com>