scripts/posix_net_conf.sh: filter out SP IRQs
Right now we know about Intel's and Broadcom's naming convention of the fast path queues vectors:
  - Intel:    <bla-bla>-TxRx-<bla-bla>
  - Broadcom: <bla-bla>-fp-<bla-bla>

So, we will try to filter the etries in /proc/interrupts for IRQs we've got from get_all_irqs_one()
according to the patterns above.

If as a result all IRQs are filtered out (if there are no IRQs with the names from the patterns above) then
this means that the given NIC uses a different IRQs naming pattern. In this case we won't filter any IRQ.

Otherwise, we will use only IRQs which names fit one of the patterns above.

Signed-off-by: Vlad Zolotarov <vladz@scylladb.com>