lowres_clock: drop cache-line alignment for _timer
Version 7 of GCC adds support for over-aligned new which also means that
it has gained awarness about the code incorrectly using plain new for
allocating memory for objects with alignent requirements larger than
max_align_t.

In case of lowres_clock the problematic member was _timer which was
aligned to a cache line apparently to avoid false sharing. However,
that should not really be necessary as it is the only non-static member
of the class. There is, indeed, lowres_clock::_now which is read by
all CPUs, but it is appropriately aligned and, since it is static,
is going to be far away from _timer.

Message-Id: <20170131195730.5068-3-pdziepak@scylladb.com>