detector for tasks blocked
Due to our cooperative model, if a task is buggy and end up going into
an infinite loop or otherwise blocking for an excessive amount of time,
we lose the reactor - potentially forever. In the absence of metrics
those problems can be really hard to debug.

This patch brings back the old task_quota signal that is now unused
since we moved that to a thread, but only fires the signal once the task
quota thread detects that we have been stalling for too long. This will
print a backtrace that will at least allow us to see what happened.

This mechanism can also be easily adapted to detect tasks that have not
respected their task quota quantum by lowering the stall threshold,
which can be done with a command line parameter.

Signed-off-by: Glauber Costa <glauber@scylladb.com>