future-util: introduce when_all_succeed()
when_all() returns a tuple or vector of resolved futures and it is the
user code responsibility to correctly handle any potential exceptions.

This is often inconvenenient as each future needs to be checked for
errors and in case there is one the remaining futures need to be
explicitly ignored.

Since, many users are interested just in the actual returned values (or
just in the fact that the future resolved) and if there are some errors
they don't care which one it is and whence it came a new construct is
introduced: when_all_succeed().

when_all_succeed() behaves very much like when_all() except that it also
checks whether there were any errors. If there were a failed future
containing one of the exceptions is returned. Otherwise, the future
(or vector) of futures is transformed by removing valueless futures,
extracting the values from the remaining ones and if any of the value
is a tuple containing a single object it is unpacked.