ninja: unbreak "ninja -t clean"
After the recent addition of c-ares into the build, "ninja -t clean"
left the build directory in an unusable state - a further "ninja" will
not be able to build Seastar.

The problem is that "ninja -t clean" only deletes targets it knows about,
not the entire build/ directory. Specifically, the target we wanted from
c-ares was build/release/c-ares/.libs/libcares.a, so this was the only
file removed. The problem is that if we remove this file and run c-ares's
"make" again, NOTHING gets built. This is an unfortunate consequence of
using libtool - the actual library is considered a byproduct, and what
the Makefile thinks the real target is build/release/c-ares/libcares.la
(a small text file). So to force .libs/libcares.a to be rebuilt, we need
to make sure that libcares.la is deleted by "ninja -t clean". For that,
we need to specify it as a target.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <20170202170405.13462-1-nyh@scylladb.com>