ninja: unbreak build ordering
Our linking stage now depends on libcares-seastar.a to exist, but it
did not specify this as an explict dependency. Rather, we have a generic
dependency of the entire build on this library, but we don't know *when*
it will be built. Lucky users got this library built before it was needed
and compilation worked, but unlucky users (like me) tried to link with
this library before it was built.

So this patch adds a dependency in the linking rules to require that
the c-ares library is built before linking. Now "ninja" should work
even for unlucky users.

The implementation adds a general concept of $built_libs, targets which
we need to actually build before we can use $libs.
Right now the only library here is libcares-seastar.a, but in the future
we can also move the dpdk library targets to this new facility (we have
an ad-hoc treatment for those in dpdk_deps, which continues to work as
before).

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <20170202171509.4936-1-nyh@scylladb.com>