tutorial: pick the name "asynchronous function".
Previously I mentioned both names "asynchrnous function" and "promise"
as two valid names for the same thing. But the term "promise" is confusing:
Different future-based implementations use it for different things, and
we also have "promise" variables. So instead, explain the term "promise"
and mention we will avoid it. Throughout the tutorial, we'll prefer the
term "asynchronous function". We already do.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <20170205151157.22426-1-nyh@scylladb.com>