print_safe.hh: factor out code that converts integer representations
Done so we have the option of converting to a buffer or directly to
stdout.  The code to convert integers uses a temporary buffer, since we
want the buffer to be filled from left to right. It is possible to fill
the buffer from the left directly without the intermediate buffer but
the code is massively simpler this way.

Signed-off-by: Glauber Costa <glauber@scylladb.com>