backtrace: prints to a temporary buffer, and then to stderr.
If various shards print the backtrace at the same time, the resulting
backtraces are unreadable. Fix that by accumulating the result in a
buffer first and then pushing the whole result into stderr at once.

Theoretically it can still happen that the output will be garbled since
the buffer has a finite size. But 8kB should be enough for most
backtraces.

Signed-off-by: Glauber Costa <glauber@scylladb.com>