Concepts support
Add experimental support for C++ concepts.  Concepts are wrapped with
a GCC6_CONCEPT() macro to hide them when concept support is disabled.

A few methods in future<>, as well as some primitives in future-util.hg are
now constrained.

Note that concepts support is experimental and may change until the
Technical Specification is approved.

Message-Id: <20170213182317.29082-1-avi@scylladb.com>