core/fstream: Avoid issuing read requests beyond _remain
Currently we issue read_dma_bulk() requests which may read past
_remain due to align_up(). If _pos + _remain is at the file end, this
will result in two io_submit() requests, because the first one will
not satisfy the request and after the second one we discover EOF.

If we don't make requests past _remain, only one io_submit() will be
issued. It's fine that the end position is unaligned, read_dma_bulk()
is aligning it internally.

Because read_dma_bulk() may reutrn more data than we requested, we
need to check if we need to trim now. Before we were fine without the
check because end was always aligned, in which case read_dma_bulk()
returns exactly as requested.

Message-Id: <1487786947-29628-1-git-send-email-tgrabiec@scylladb.com>