Add metrics-hostname to the configuration
It is sometimes useful to set the hostname used by metrics
implementation insted of using the machine hostname, either because it
would have more meaning, or when running multiple instances on the same
host.

This patch Adds a configuration parameter set it in app_template and
modify the prometheus and collectd implementation to use it.

See scylladb/scylla#2110

Signed-off-by: Amnon Heiman <amnon@scylladb.com>