prometheus: Http handler that returns prometheus 0.4 protobuf or text format
The prometheus binary protocol is the efficient way for sending prometheus
data. It is many times helpful to look at the metrics directly at the
server, for that the text protocol of the API is more convinient.

This patch change the prometheus API to respect the accept header. When
the client accept protobuf, the results will be in the binary format, if
not, the results will be in the text format.

Fixes #221

Signed-off-by: Amnon Heiman <amnon@scylladb.com>
Message-Id: <1488988809-20509-2-git-send-email-amnon@scylladb.com>