Fix smp::submit_to() with function reference
smp::submit_to() tries to cleverly optimize the case of submitting a
functor to run on the current CPU and the functor is an lvalue, in which
case we can avoid increasing its lifetime a la do_with() - which we do
need to do in case the functor is an rvalue.

The problem is that all that clever code does not correctly handle the
case of a normal C function, e.g.,

	extern future<> f();
	return smp::submit_to(c, f);

In this case we get a reference to a function, which cannot be used
in make_unique, and causes a compilation error.

The solution in this patch is to use std::decay_t<> to convert the
function reference to a traditional function *pointer*. This pointer
can be copied around. It's perhaps not the most efficient solution (we
allocate memory to copy this pointer), but any other solution would be
much more complex and in any case, won't optimize the submit-to-different-
cpu case in which we (currently) always allocate.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <20170312150329.13981-1-nyh@scylladb.com>