tutorial: make network examples work on multi-core
So far, the network server examples explained that they only work on
a single core ("-c1"). This was unduly confusing for beginners, who read
this limitation and then promptly forgot all about it :-)

So better just show, up-front, how to run the service loop on all the cores
in the simplest way, using smp::submit_to(). Then all the network server
examples will work with any number of cores.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <20170314143234.20327-1-nyh@scylladb.com>