Metrics: Collect metrics under the same metric name
This patch address the following issues:
* Collect all metrics with the same name (but different labels) under
the same map entry.
* Verify that a metric will be only register once
* Verify that a metric has a single type
* Adds to the value data structure a shared pointer with the id and
description, which makes it easier for the higher layer (prometheus and
collectd) but without the penalty of copying it.

The computation assumption is that while metrics rarely change during a
normal run, they are often read.

Schematically, the registration would look like:
map<name, metric_family >

Where metric_family is a map<labels, registration> with some additional
information about the metric. Currently only the type is there, but
additional metric information will be added in following commits.

Name is the metric name ({group}_{name})
And labels is a map of {label name, label value}.

The registration object holds the id information.

When getting a copy of the values it will be:
map<name, vector<{registration, value}> >

The registration is passed as a shared pointer and holds any information
needed by the upper layer.