Use safe name for metric family
Metric family name is used by prometheus to report the metrics under it.
As such, it should be safe (no funny character like minus or space).

There is no need to do this safe manipulation for each protocol on each
metrics collecction, instead, we can use a safe name for the metric
family.

Fixes #269

Message-Id: <20170330091758.18535-1-amnon@scylladb.com>