fstream: do not shrink buffer size if unused/total ratio is below target
update_history_consumed() attempts to increase the buffer size as much
as possible so that the unused/total ratio stays below the target even
if the next buffer is fully discarded.
If that happens set_new_buffer_size() should not attempt to decrease the
buffer size (provided that the ratio is still good) to avoid these two
parts of logic fighting with each other.