perftune: remove psutil dependency
Because psutil is not available in some major distributions - like
Centos7 - as a python3 package, we have to go to great lengths to
provide the additional package to support this in a clean way.

While that is okay, psutil really is only used for a dirscan that can be
done in a couple of lines of code. It really isn't worth it to pull an
extra dependency just for that.

This patch rewrites the two methods that we need and use them.

Signed-off-by: Glauber Costa <glauber@scylladb.com>
Message-Id: <20170406140653.3015-1-glauber@scylladb.com>