scripts: perftune.py: fix the compute CPU mask calculation for NVMe-only case
If disks that need to be tuned are only NVMe disks - the tuning mode is a MQ mode
when disk's IRQs are equally distributed among all available CPUs and compute
CPUs are all available CPUs.

This in particular means that in this case the configuration mode doesn't depend on the available
CPUs configuration.

Signed-off-by: Vlad Zolotarov <vladz@scylladb.com>