scripts: perftune.py: use a 'ps' command instead of scanning system files
The current code has a main weakness in the is_process_running() implementation:
reading the PIDs and their further analysis are two separate stages.

This allows some processes (and their PIDs) to disappear between these two stages.

This patch eliminates this problem by removing these two stages by using `ps -C <cmd name>`
command that does all we need.

We adjust to the run_one_command() to be able to ignore the non-zero exit status
and simply run the 'ps --no-headers -C <cmd name>' which would either return the corresponding
lines if such process runs or will exit with a non-zero exit code if there are no running processes
with a given name.

We also filter out the zombie processes that may be a part of the 'ps' output.

Fixes #281

Signed-off-by: Vlad Zolotarov <vladz@scylladb.com>