scripts: perftune.py: use a pyudev 0.16 API
Fedora and CentOS7.2 come with the latest 0.21 pyudev version while Ubuntu 16.04 comes
with an older 0.16 version.

This patch changes the code to use the older API with works with both 0.16 and 0.21 module versions.

In particular this patch changes the code to call the pyudev.Device.xxx() instead of  pyudev.Devices.xxx().
It also adds the missing error check that was found while debugging this issue.

Fixes #280

Signed-off-by: Vlad Zolotarov <vladz@scylladb.com>