add special logger for memory allocation failures
It is often hard to tell what is going on when system start getting
allocation failures. Is it due to real memory shortage or because of
large memory allocation and memory fragmentation. Or may be it is
because of a bug in a code that asks for to much memory.

This patch adds a memory logger that logs useful diagnostic information
on allocation failure and can be enabled dynamically when system start
to malfunction.

The patch logs allocation size, current stack trace and memory statistics.

The caveat is that logging itself may fail to allocate, but the worst
that can happen is that logging will fail, the function is protected
against recursion.

Message-Id: <20170528130009.GB6518@scylladb.com>