core: Add `lowres_system_clock`
This change is motivated by scylladb/scylla#1957.

The existing `lowres_clock` is based on `std::chrono::steady_clock` and produces
time points relative to an opaque epoch. `lowres_system_clock` has the same
precision as `lowres_clock`, but is based on `std::chrono::system_clock` and has
a fixed epoch as per the usual Unix time.

Since both clocks have the same precision, a single timer updates the time
points for both clocks.

While strictly speaking `std::chrono::system_clock` is not required to use the
Unix time epoch, this is a defacto convention that is commonly
assumed (including in Scylla).

Testing: All unit tests pass. The unit tests for scylladb/scylla pass when it is
built against Seastar with this patch.