Http: Reply to support outstream
For long messages it is some times useful to use chunked in the reply.

In this mode, the handler uses an output stream instead of returning a
string.

The output stream is used only for writing the message body, it would
break the body into chunked.
Each call to the buffer write is also an oportunity to yield, so the
messages handler will not block the shard.

This patch change the reply message, so it can create an http body
handling stream.

The user can call write_body with a function that accept an output
stream and return a future.

Note that at this point the server would return 200 OK as part of the
headers. So do not use it before performing all parameter validation.