prometheus: Use output stream to return the metrics
Generating the metrics reply can take a long time, to reduce the load on
the system the prometheus can use the output stream api.

This patch changes the message creation to use the output stream.
This both reduce the buffer size needed and make the prometheus yield so
it will not block the shard.

The patch is partially indented on purpose and an indentation patch will
follow.

See scylladb/scylla#2475

Signed-off-by: Amnon Heiman <amnon@scylladb.com>