core/thread: Fix unwind information for seastar threads
There is no caller of main(), so information unwinders get for the
frame of main() is incorrect, it points to some garbage.

Because of that, the unwnider used by the perf tool dropped backtrace
for any code running in a seastar thread.

Fix by adding a CFI entry (dwarf) marking IP as undefined, just like glibc
does for entry frames, e.g. in clone().

See scylladb/scylla#1909

Message-Id: <1499077464-16538-1-git-send-email-tgrabiec@scylladb.com>