tls: Wrap all IO in semaphore
Fixes (hopefully) issue with assert in vec_push @ shutdown

Basic problem is that to make GNUTLS:s aync IO work in our
version of async, we need to keep a "fake" EINTR state machine
of pretending not to have yet gotten/sent data, while in fact
we are actually _doing_ so, just not ready.

This is fine, assuming the usual rules of max 1 reader + 1
writer at a time. However, we also have a semantic saying
"shutdown" (or output_stream::close etc) interrupts a pending
IO op (i.e. read). And TLS is not like all other chickens,
we want to keep both the engine and other endpoint happy, by
sending handshakes for termination. This actually means we
need overlapping IO read/writes of sort.

This leads to a race, where we sometimes cannot be sure
which data op is actually pending. We could maybe build a
complex data queue (chain futures or whatnot). But that
sounds nightmarish.

Now, looking at how the actual handshake is done in gnutls,
we can probably get away by simply keeping semaphores for
read&write, and enforce using these _around_ read/write
and corresponding shutdown handshakes.

When doing shutdown(WR), we only need to be able to send
an alert across the wire, so assuming writes will complete
(which is a big assumption though - here is a huge dragon)
we can ensure we always do either shutdown case by
first initiating session termination and hope is will
terminate gracefully quickly. Otherwise, the timer+underlying
close will do it less so.

Message-Id: <1499789341-18635-1-git-send-email-calle@scylladb.com>