HTTP: httpd support output stream in reply
Now that reply can have an output stream, the http server check when
handling the reply if it was sent by the handler, if so, it should use
the registered function to create the reply body.

In case of a failure, it would throw all messages from the queue and
close the connection.

Signed-off-by: Amnon Heiman <amnon@scylladb.com>