Add overall python code QA and guidelines with flake8
Seastar loves python &	python loves Seastar.

It would benefit the project to	start enforcing	some code guidelines
and basic QA with a linter along a PEP8 respect	thanks to flake8.

This patch adds	a tox config to	at least start with an assessment
of the work to be done on all .py files	in the code base.

To reduce its noise, tests on long lines (> 80char) are ignored
for now.