Provide workaround for non scaleable c++ exception runtime
g++/glibc combination provides a c++ runtime with non scaleable
c++ exception. This is dues to global locks in stack unwinding code.

To summarize all the locks we have now and their purpose:
 1. There is a lock in _Unwind_Find_FDE (libgcc) that protects
    list of "Frame description entries" registered with __register_frame*
    functions.  The catch is that dynamically linked binary do not do that,
    so all that it protects is checking that a certain list is empty.
    This lock no longer relevant in gcc7 since there is a patch there
    checks that list is empty outside of the lock.
 2. The lock in dl_iterate_phdr (glibc) that protects loaded object
    list against runtime object loading/unloading.

To get rid of the first lock one has to use gcc7.

To get rid of the second one we can use the fact that we do not
load/unload objects dynamically (at least for now). To do that we
can mirror all elf header information in seastar and provide our
own dl_iterate_phdr symbol which uses this mirror without locking.

Unfortunately there is another gotcha in this approach: dl_iterate_phdr
supplied by glibc never calls more then one callback simultaneously as an
unintended consequences of the lock there, but unfortunately libgcc relies
on that to maintain small cache of translations. The access to the cache is
not protected by any lock since up until now only one callback could have
run at a time. But luckily libgcc cannot use the cache if older version
of dl_phdr_info is provided to the callback because the older version
did not have an indication that loaded object list may have changed,
so libgcc does not know when cache should be invalidated and disables it
entirely. By calling the callback with old version of dl_phdr_info from
our dl_iterate_phdr we can effectively make libgcc callback thread safe.

Message-Id: <20170727123008.GB2684@scylladb.com>