json object to support output stream
The http server now support writing body to an output stream.

If the return json object is very large, it would be good to use the
ouptput stream instead of serializing to string.

This patch adds support to streaming a json object.

The json_return_type can now hold a serialization function that would be
pass to the reply instead of a string.

To use it, instead of creating the the json_return_type from the object,
use the stream_object helper function.

return make_ready_future<json::json_return_type>(stream_object(res));

stream_object return a function that captures the object and will be
used to serialize it to the output stream.

The formatter now supports write method that serialize an object to
an ouput  stream.