scripts: perftune.py: fix the CPU mask generation for more than 64 CPUs
For setups with more than 64 CPUs hwloc-distrib generates funny masks like
"0x00ffffff,,,0x0" where empty spaces between comas represent the zero mask.

Unfortunately /proc and /sys files we configure don't like this form of a mask
and want the zeros to be actually present. For the example above they need to be
given the following mask: 00ffffff,0,0,0.

If the mask is not in the expected form the modification simply fails.

This patch makes sure that we mask is as /proc and /sys expect it to be.

Fixes #316

Signed-off-by: Vlad Zolotarov <vladz@scylladb.com>
Message-Id: <1501776960-6816-1-git-send-email-vladz@scylladb.com>