semaphore: add a version of with_semaphore that takes a duration timeout
Waiting with timeouts is a very common operation. with_timeout is a nice
wrapper that is in some places where we just need to execute a function
more elegant than get_units, because we don't need to play with manually
passing the permit to continuations inside the function.

with_semaphore calls get_units, and get_units currently only have a
time_point, not a duration parameter. We'll provide this as well, as
both variants are commonly found in callers.

Signed-off-by: Glauber Costa <glauber@scylladb.com>
Message-Id: <1501895628-22986-1-git-send-email-glauber@scylladb.com>