Provide a convenient way to lazy-convert to string the values of pointers
One might want to log the value of *p but this introduces a problem. If
p is null this will crash, so instead something like
    p ? to_string(*p) : "null"
has to be used. This will however introduce a potentially unnecessary
branch when used to produce a msg for trace() or debug().
lazy_deref(p) will defer this check to when the value is actually
converted to string.

Signed-off-by: Botond Dénes <bdenes@scylladb.com>
Message-Id: <ba723c687e4b7dc9490d3b047dfe3b8bb03a6ae8.1502120671.git.bdenes@scylladb.com>