tls: Only recurse once in shutdown code
Fixes #321

"shutdown" needs to ensure we can take semaphore locks, and in correct order.
It does so via recursion in the shutdown_wr case. However, it is timing
dependent whether the recursion condition is cleared on next lap. And we
can wind up SOE:ing.

Should be back merged to whereever tls is ever shut down. At least
in contended conditions.

Message-Id: <20170814113825.7604-1-calle@scylladb.com>