Http test: Test json using output stream
This patch adds a test that serialize an object that result in a very
large json.

It adds a test helper function run_test, that gets two functions, on
that write to a strem, and a second that read from a stream.
To prevent issues with the client, the http parser can now work in count
mode only, it would parse the body and count its size, but will not hold
the entire body as a single string.

A helper class mimic a json object that would be serialized to a
1,000,000 bytes json string.

A vector with 1,000 entries is created with this object.
And the code verify that this vector can be passed from the server to
the client.