util: Rename logging structures
The name of the `log_registry` class was differentiated from the global logger
registry singleton (`logger_registry`), but this is confusing.

Instead, use the logical name for the class and indicate the singleton with
`global_logger_registry`.

This should introduce relatively little churn, since few callers access these
names directly (and Scylla accesses them in an abstraction layer).