memory: add detector for large allocations
Since large allocations can fail or cause cache eviction in the face
of memory fragmentation, applications should try to limit allocation sizes.
This patch adds a detector to assist application developers in finding
call sites of large allocations.

A facility is provided to disable it temporarily when a large allocation
failure can be tolerated.

Message-Id: <20170819135006.24993-1-avi@scylladb.com>