rpc: make sure that _write_buf stream is always properly closed
Currently _write_buf is closed when send loop exists, but if an error
happens (or a client is closed deliberately) during negotiation,
before send loop is started, _write_buf is never closed although
it was used for sending negotiation frame, so it has to be flushed
and closed() otherwise "Assertion `!_in_batch' failed" may happen.
Fix that by moving stream closing from the end of send loop into
the code that is always called during client shutdown.

Message-Id: <20170919141225.GG32385@scylladb.com>