Work around GCC 5 bug: scylladb/seastar#338, scylladb/seastar#339
This fix prevents running into std::tuple construction failure when the constructor of a stored type expects a unique_ptr to the base class but a unique_ptr to a derived class is passed.
The original code compiles fine with GCC 6 and later so can be restored once Seastar migrates entirely to newer compilers.

Fixes #338

Signed-off-by: Vladimir Krivopalov <vladimir.krivopalov@gmail.com>
Message-Id: <20170925044647.23450-1-vladimir.krivopalov@gmail.com>