reactor: preempt tasks when a signal is received
When a signal is received, preempt running tasks so the signal can be
processed quickly. This is important for the high-precision timer used
by threads with a thread_scheduling_group to set their next wakeup
point.

Fixes test_thread_scheduling_group. It's not clear why the scheduler
broke it, it could be that an extraneous poll was removed, which triggered
the problem. In any case, thread_scheduling_groups will be deprecated soon.
Message-Id: <20171003083854.9451-1-avi@scylladb.com>